fix: move success criteria to seperate class and consolidate builders
